Renamed Configurator.js to configurator.js in src/core and test/core Updated references where relevant
