+++ src/core/configurator.js
