+++ src/gladius-src.js
@@ -5,7 +5,7 @@ define( function ( require ) {
     var lang = require( './core/lang' ),
         _Math = require( 'math/math-require' ),
 
-        Configurator = require( './core/Configurator' ),
+        Configurator = require( './core/configurator' ),
         ThreadPool = require( './core/threading/pool' ),
         Scheduler = require( './core/scheduler' ),
         Event = require( './core/event' ),
