+++ test/index.html
@@ -16,7 +16,7 @@
     <script src="core/entity.js"></script>
     <script src="core/component.js"></script>
     <script src="core/scene.js"></script>
-    <script src="core/Configurator.js"></script>
+    <script src="core/configurator.js"></script>
     <script src="core/scheduler.js"></script>
     <script src="core/component/transform.js"></script>
     <script src="core/resource/script.js"></script>
